+++ packages/react-querybuilder/src/utils/parseNumber.ts
@@ -1,5 +1,5 @@
+import { numericQuantity } from 'numeric-quantity';
 import type { ParseNumbersMethod } from '../types/index.noReact';
-import { numericRegex } from './misc';
 
 /**
  * Options object for {@link parseNumber}.
@@ -9,12 +9,24 @@ export interface ParseNumberOptions {
 }
 
 /**
- * Converts a string to a number, or returns the string unchanged if it does not
- * match {@link numericRegex}.
+ * Converts a string to a number. Uses native `parseFloat` if `parseNumbers` is "native",
+ * otherwise uses [`numeric-quantity`](https://jakeboone02.github.io/numeric-quantity/).
+ * If that returns `NaN`, the string is returned unchanged. Numeric values are returned
+ * as-is regardless of the `parseNumbers` option.
  */
 export const parseNumber = (v: any, { parseNumbers }: ParseNumberOptions) => {
-  if (typeof v === 'bigint' || typeof v === 'number') {
+  if (!parseNumbers || typeof v === 'bigint' || typeof v === 'number') {
     return v;
   }
-  return parseNumbers && (parseNumbers === 'native' || numericRegex.test(v)) ? parseFloat(v) : v;
+  if (parseNumbers === 'native') {
+    return parseFloat(v);
+  }
+  const n =
+    // TODO: Should these options be configurable?
+    numericQuantity(v, {
+      allowTrailingInvalid: parseNumbers === 'enhanced',
+      romanNumerals: true,
+      round: false,
+    });
+  return isNaN(n) ? v : n;
 };